+++ test/unit/test-utils.ts
@@ -27,8 +27,6 @@ afterEach(function () {
     clearTimeout(stage.dblTimeout);
   });
 
-  console.log(isFailed);
-
   if (!isFailed && !isManual) {
     Konva.stages.forEach(function (stage) {
       stage.destroy();